requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0
environs==9.5.0
exceptiongroup==1.1.3
fastapi==0.104.0
fastapi-mail==1.4.1
fastapi_cors==0.0.6
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gspread==5.12.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
passlib==1.7.4
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.23.2
